+++ packages/render-composer/src/index.ts
@@ -3,10 +3,10 @@ export * from "./CopyPass"
 export * from "./DepthCopyPass"
 export * from "./EffectComposer"
 export * from "./EffectPass"
-export * from "./PreRenderPass"
+export * from "./LayerRenderPass"
+export * from "./Layers"
 export * from "./RenderPass"
 export * from "./RenderPipeline"
 export * from "./SelectiveBloomEffect"
 export * from "./SMAAEffect"
 export * from "./VignetteEffect"
-export * from "./Layers"
